+++ RuneFleet.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.14.36212.18 d17.14
+VisualStudioVersion = 17.14.36212.18
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OSRSClientManager", "OSRSClientManager\OSRSClientManager.csproj", "{348F904D-1F84-49D9-B048-2E8D11DCF8A6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RuneFleet", "RuneFleet\RuneFleet.csproj", "{348F904D-1F84-49D9-B048-2E8D11DCF8A6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
